Manage config validation code on its own
